Moved /generate to new route in common_services.py
