+++ tests/test_noq.py
@@ -12,7 +12,12 @@ def test_users():
 
 
 def test_create_user():
-    data = {"name": "Victoria","phone": "0708504033","email": "victoria@queen.se","unokod": "123"}
+    data = {
+        "name": "Victoria",
+        "phone": "0708504033",
+        "email": "victoria@queen.se",
+        "unokod": "123",
+    }
     response = client.post("/users", json=data)
     assert response.status_code == 200
 
@@ -29,3 +34,8 @@ def test_reservations():
     assert response.status_code == 200
     json_data = dict(response.json()[0])
     assert int(json_data["user_id"]) > 0
+
+
+def test_generate():
+    response = client.get("/generate")
+    assert response.status_code == 200
